Fix python-libjuju CI test issues
Juju is having trouble bootstrapping on lxd 5.0 due to autoload
credential issues (bug is being filed), so this lets the tests use a
(more) stable version for the time being. Also disabling apparmor
prevents juju to connect to the newly created controller container,
thereby causing bootstrap to hang (not sure why atm -- not doing so
seems to make it work).