Build dqlite static libraries
The following builds the dqlite static libraries that can be downloaded
when compiling the dqlite branch in juju/juju. This enables the
developers of juju to not have to be worried about setting up a build
farm for doing this.

We're still missing a lot of the architectures, but we can still move
towards that in the future.

For now only amd64 and arm64 is being created here. This should allow us
to start testing the dqlite branch on graviton in aws for example.