Enable fuzz checking on amd64 for 30s
The following enables fuzz checking on amd64 for 30s to test the
various scenarios. The more fuzzing we add the slower it will
become. This will just test if it's actually useful. We could
move this to it's own job for fuzzing only, that way we're not blowing
out other jobs.