Setup parca-agent for ci-runs
Start collecting profiling data for ci-runs (specifically looking
at longer gating test runs). We want the data for 2 reasons:

 1. See if there are any potential bottlenecks in terms of performance
    that we can see when the jujud and jujud-controller is running.
 2. We can use profile guided optimisations (PGO) from the runs to
    apply to the binaries.

This currently only applies to lxd deployments. For other providers
we will need to install it via cloud-init or bake it into an image.

This is an experiment to see if it's useful or not.