test(microk8s-version): bump microk8s version to 1.34
bump microk8s version to 1.34 because Ubuntu 24.04 and higher have introduced a change in apparmor
rules that broke the strict microk8s snap. See https://github.com/canonical/microk8s/issues/5190.
This problem shouldn't not be present with non strict snaps, which we are using here. However the
failures I am seeing are very similar to ones related to AppArmon (juju fails to even create the
namespace to bootstrap the controller). So this is a test to see if 1.34 will solve the issue.